refactor(agents): move the requester settle wake into its own module [AI]
Two pieces of fallout from rebasing across upstream tooling changes:

- The max-lines lint budget (#107315) has no grandfathered suppression for
  subagent-announce.ts, and the settle wake + retired-row ledger pushed it
  past the cap. The wake is a coherent unit, so it moves to
  subagent-announce.requester-settle-wake.ts (matching its test file) with
  its own registry-runtime seam; the registry's lazy dep and the lifecycle
  type alias point at the new module. announce.ts drops back well under
  the budget with no suppression needed.

- The announce read-path refactor rewired the runtime barrels onto
  subagent-registry-announce-read.js, orphaning the
  hasDescendantRunAwaitingSettle re-export on subagent-registry.ts, which
  the deadcode-exports gate now flags. Removed; the retry-grace e2e reads
  the announce-read implementation directly.
