+++ src/agents/subagent-registry-lifecycle.ts
@@ -80,7 +80,7 @@ type CaptureSubagentCompletionReply =
   (typeof import("./subagent-announce.js"))["captureSubagentCompletionReply"];
 type RunSubagentAnnounceFlow = (typeof import("./subagent-announce.js"))["runSubagentAnnounceFlow"];
 type MaybeWakeRequesterAfterAllChildrenSettled =
-  (typeof import("./subagent-announce.js"))["maybeWakeRequesterAfterAllChildrenSettled"];
+  (typeof import("./subagent-announce.requester-settle-wake.js"))["maybeWakeRequesterAfterAllChildrenSettled"];
 type BrowserCleanupModule = Pick<
   typeof import("../browser-lifecycle-cleanup.js"),
   "cleanupBrowserSessionsForLifecycleEnd"
